codes/requirements.txt
streamlit
langchain_community
langchain_openai
langchain_core
openai
python-dotenv
faiss-cpu
pandas
fastapi
uvicorn